// src/app/modules/master-forms/Models/companythreshold.model.ts
export class Companythreshold {
  BranchId: number = 0;
  BranchName: string = '';
  CompanyId: number = 0
  ThresholdValue: number = 0;
  RaiseClaimDay: number = 0;
  ClaimSettlementDay: number = 0;
  CompanyCode: string = '';
  CompanyName: string = '';
  InStateAmt: number = 0;
  OutStateAmt: number = 0;
  SaleSettlePeriod: number = 0;
  NonSaleSettlePeriod: number = 0;
  Action: string = "";
  PkId: number;
  addedby: string = '';
  addedOn: string = '';
  lastUpdatedOn: string = '';
}